+++ src/pages/DepositUSD.tsx
@@ -94,7 +94,6 @@ const testTokensData = [
 
 const selected = {
   maxSlippage: 0.1,
-  gas: "standard",
 }
 
 const testTransInfoData = {
